relocating the debugbar throwable marker to the contracts namespace
